Add Stable and Versioned Publishing Job
Triggered when a tag in the form `v<semver>` is applied by someone in
the core team.

Sets up `/stable` and `/v<semver_withoutperiod>` (ex: `v41` for tag
`v4.1`).

This PR also adds the ancillary files for v41 and stable. In the future
ancillary files will have to be added as one of the last PRs before
tagging occurs (like `www/v41` included here).

Note one of the steps in the versioned work is to set `stable_version`
in `_config.yml`, this variable is currently un-used but we may have use
for it in the future for linking or banner control.

Please review this job thoroughly so that we can avoid the misses we had
on the 'latest' deploy job :wink:

Signed-off-by: kingthorin <kingthorin@users.noreply.github.com>