improve: GitHub Action Workflows no longer include file ext on links
Instead of switching .md to .html just drop the extension entirely.

GitHub pages seems to serve .md, .html, and no extension normally so
even if someone happens to have linked content with .html then it'll be
fine.

Changed both the "latest" job and the "stable" job, existing content for
v4.1 and v4.2 will continue to have the html extensions, but that's
fine.

Signed-off-by: kingthorin <kingthorin@users.noreply.github.com>