Introduction - Add "Disable Compensating Controls For Testers" (#1131)
* Introduction - Add "Disable Compensating Controls For Testers"

Add a header to the introduction page recommending disabling compensating controls such as a WAF. Testers can use this section to provide product teams with an authoritative resource on how to set up a penetration test effectively.

There is a widespread belief that penetration tests should be conducted in a black-box setting as this is more realistic. While true, the goal of this PR is to help educate that white-box penetration tests without compensating controls are much more effective in securing an application.